Added IOMan::implRemove(const IOObj&) function overload [devel]
Summary:
This function can be used with a temp IOObj that does not have an entry (yet) in the omf file. Even otherwise, it avoids fetching the IOObj twice.
Also implemented SegyTrcTranslator::implRemove to make sure Segy files are never deleted by OD I/O system.

Test Plan: exported SEGY files in batch and non-batch modes. No more crash when the tempID entry is missing from omf. Also, the segy file is not deleted when there is a tempID entry in the omf for another SEGY file.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14083