Fixed error in odpy.common.getODSoftwareDir
If the input dictionary is not null and contains the
key 'dtectexec', four levels of the corresponding file path
need to be removed in order to really return the root of the
OpendTect software installation, and not the path to the
binaries.

Also improved odpy.common.getODArgs(), so it reads
the survey location from the setting files of the current
user.

[Jira: OD-3185 and OD-3225]