Fixed interpolation of well displays
Summary:
Fixing in a different way the issue OD-3120, to overcome possible
shifts issued by this patch, as reported in OD-3207.
Only impacts well tracks which are truncated: the first MD is a rather large,
non-zero value, or the last track MD value is lower than the actual TD of the well.

First I reverted the extrapolation of the Time-Depth model in the function
getDatumTwtFromFile. Needs to stay off to avoid introducing bulk time shifts
as seen by Paul Glenn.
Secondly the implementation of Well::Track::getPos is updated to enable
getting a good value, linearly interpolated in 3D and not only vertically,
if the argument is outside of the track range.
This value is then used by the OpendTect well reader implementation to extend
the track is memory when logs and/or markers are found outside of the track
range.

Test Plan:
Displayed wells in many surveys, well displays are fine, in terms
of extension and absolute position.
The cross-plot tool is capable of extracting the data to the full range of the log,
thus possibly outside of the track range (on disk).

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14112