Fix Presentation Maker plugin issues with custom file template
Summary:
Fixes OD-3263
Need to make sure the user understands that a custom template just needs to be 2 slides:
- The front page where the presentation title will be shown.
- The second page which will be used for all the content slides. Each content slide has a title at the top
In both cases the first placeholder/shape on each slide will be  where the titles should be added. The
template is not a powerpoint template (ie .potx) file it is just an ordinary .pptx.

I also enabled getting the slide size from the file so that the user doen't have to mess with that to
get the screengrabs page centered.

Test Plan:
Tested on Linux:
I've made an example custom template based on the dGB standard presentation, maybe this can be added
to the documentation or as part of the software distribution for users to reference.
Slides appear OK when displayed in LibreOffice

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D14117