Fixed new windows sometimes falling behind at startup
Summary:
Made a good alternative for uiMainWin::showAlwaysOnTop, which should not
be used unless strictly necessary. This function is actually a trick to avoid
a focus issue on windows, which does not raise a window for a new application,
see the documentation of QWidget::activateWindow().
I propose here to activate the window just before starting its event loop.
Since the Qt flag to handle the window is static, and cannot be retrieved, I
introduced an enum that matches QWindowsWindowFunctions::WindowActivationBehavior,
and that is stored in a static window.

Test Plan:
Needs to be tested on Windows, in an installed project. Works nicely,
no side effect seen. However I still need to search other showAlwaysOnTop and
replace them by setActivateOnFirstShow for selected dialogs.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14120