Filled the user commands toolbar when the user settings are empty
Summary:
If the python settings for the user are empty, one still wants to
try to fill the user commands toolbar with something. As the CommandDefs
class returns an non-empty BufferStringSet is any executable is found,
using it to provide some default.
I still want to check how to safely compress a bit the logic in uiSettingsMgr::updateUserCmdToolBar,
as their is some redundancy within the if-else if-else statements.

Test Plan:
Python settings and terminal icons always appear in the toolbar
at application startup, even in the absence of stored python settings
(or link to the Python installation). Python IDE icon only appears when the Internal
python installation is registered in the user settings.

Reviewers: nanne

Reviewed By: nanne

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14126