Added functions to facilitate external bindings [devel]
when loading OpendTect libraries from external applications,
many thing may go wrong because the SetProgramArgs is not called,
the init function of each module is not called, ...
Three external C functions are thus added to facilitate it, knowing
that we may not be running an OpendTect executable, and that the current
executable location can be outside of the OD binary tree.

Tested from the python_activator stand-alone application.
Works correctly, with one exception (without apparent consequences), that
it dGBCommon is loaded, the libraries are never released. This may come
either from the plugin loading facilitiy, or because of flexnet.
Can also be potentially used for the Python and node-gyp bindings (Terranubis)