Unified launching of Terminal prompts [devel]
Summary:
There are four ways to launch a terminal:
Ctrl+T shortcut, the toolbar icon and menu item, and the
icon in the Python settings dialogs.
This work ensures that all four user actions end up in
the same implementation in uiTools. Also it no longer makes
use uiSettingsMgr::doToolBarCmd for that purpose, as the
implementation is very specific.
Finally this removes completely the usage of the
functions in SettingsAccess, obsoleted by the CommandDefs
class.

Test Plan:
Terminal starts from all four actions on Windows,
with or without valid python settings

Reviewers: raman, wayne

Reviewed By: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14150