Fixed missing files in developpers package
Summary:
Instead of doing everything from the release build,
this packs most of the required files by running the package target
on the debug build, including the class documentation.
When running the devel target on the release build, the only thing
that should happen is that it should add the object libraries (*.lib)
from the release folder on top of the existing zip file.

Test Plan:
to be done on Windows. Running the package target in a debug build
now succesfully adds all external dependencies.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D14163