Catch possible Well::Data nullptr [devel]
Summary:
Fixes OD-3332 [CR 20210801111649] uiImportLogsDlg::acceptOK(CallBacker *) [uiwelllogimpexp.cc : 220
I am unable to reproduce the reported crash on either Windows or Linux in 6.6.4 or the 6.6 branch.
The crash report suggests the Well::Data* wd must be a nullptr. Probable culprit is the pointer check
on lines 168-169 which should return from the function but doesn't.

Test Plan: Tested on Linux: but unable to reproduce crash

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14186