Fixed HiddenParam issue in OffsetValueSeries [devel]
Summary:
Found that not all HiddenParam object were deleted before closing
the application. While debugging is turned out that the constructor
was using a different instance of cursizes than the destructor.
Removing the static keyword makes the variable global and creates
one instance per template type.

Test Plan:
Checked in debugger that objects are removed in the HiddenParam destructor.
No PE when closing the application.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14192