now placed check for launching Firewall dialog [devel]
Summary:
Earlier UAC was popped even when no process was to be added. Now a check is placed and if no process needs to be added, the UAC would not be popped.
Now the condition should be placed before launching the process. (OD-3318)

Test Plan: Will be tested with installer

Reviewers: nageswara

Reviewed By: nageswara

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14209