Rework of PropertyRef related classes
Summary:
Moved the StdType enum from PropertyRef to Mnemonic,
restored the PropertyRef class like it was in od6.6,
Cleaned up extensively the API of the Mnemonic and Property* related
classes.
Aligned the usage of units: all (property) values are stored on disk and in memory
in the unit of measure referenced by their PropertyRef

Test Plan:
Basic models can be loaded from file. More test to be done.
SynthRock to follow.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14243