Cleanup for the handling of units base classes [devel]
For Mnemonics, PropertyRef, Well::Log classes.
Providing an API to access the mnemonic and unit
to return directly either a string or object depending on the
usage.
Clarified that the label must always be the unit symbol, unless
this later is empty. Hence a conversion of the label may happen
when setting it with the public function.