Added contingency for stored units in SynthRock IOPars [devel]
Summary:
Using a FileMultiString to store units saved/retrieved
in the ValueProperty class.

This patch will allow the current branch to read such files
without issue, using only the first part of the FileMultiString
to set the incoming value, and with the possibilty to parse
an incoming FileMultiString.

Added functions to the PropertyRef class to make it clear what the
software is using internally. This also allows parsing incoming
IOPars which might be written in different units.

Test Plan: Saved simulations in SynthRock can be opened.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14285