Fixed crash during PDF creation
Summary:
the cholesky_ class member is not initialized to null
in the copy constructor. As a result the copy operator may crash
while deleting an uninitialized pointer.
I cannot change the implementation of the copy constructor as it is
inline, thus the only option is to use the other constructor in
uiEditProbDenFunc, and then copy the content.
I tried to make either the copy constructor or the clone function as
deprecated, to stress that it is unsafe, but the deprecations where not
flagged by the compiler.

[Jira: OD-3406]

Test Plan: No more crash in debug mode. Was hardly crashing in release mode

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14300