Improvements in Import 3D Horizon behavior
Summary:
No more implicit scaling of Z values or Inl/Crl <-> X/Y conversions. If the user selects wrong format definition and the scanner detects it, a warning message pops up that gives the user a chance to change the format definition.
OD-3040

Test Plan: Imported a horizon using different format definitions. The messages are relevant and the results of scan and import are correct.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14307