Fixed several issues in well data manangement
The OD implementation of the well info reader was reading the
track, but leaving the dahrange_ of the track not initialized.

Furthermore Well::Data::loadedState() was setting Info to true
when the track was loaded. There is no reason to do that, it can
be false when other implementations than the OD well reader.

Finally reworked the InfoCollector implementation, to make it
use the new API of Well::Man, with its loading granularity.

[Jira: OD-3455]