Lazy cleanup of Well::Data objects [devel]
Summary:
Added a function in Well::Man to empty the
manager if no one else than the manager is using the
stored Well::Data objects. This allows for instance releasing
the objects after removing all well tree elements if that was the
only operation done with wells, or when a plugin like SynthRock
stops using well data.

Test Plan:
tested various combinations: Wells used only in the tree,
only in SynthRock, in both. Checked in debug mode that Well::Data::wells_
is empty when the tree no longer contains well items and SynthRock
is closed, and that no other Well::Data object gets destructed until
the application is closed, without having the change project.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14362