Large rework of Math::Formula and its interface
Summary:
added support for storing mnemonics in the Math::Formula
class and its interfaces, provide API for providing the input/output
units of the values that are provided to the form, and are calculated
using it. This eliminates all need of conversions when providing data
to the form, or getting data from it.
One can provide a Mnemonic selection to uiMathFormula, which gives a
meaning to the names provided with uiMathFormula::setNonSpecInputs.
That allows moving a lot of external code into uiMathFormula,
especially it makes it possible to guess from the expression what
kind of input corresponds to variables, and set the formula units
accordingly.

Test Plan:
tested the form from the Log create dialog (with/without
rock physics), the math attribute, the math gui inside SynthRock,
and briefly the interface of the HorizonMath. The interface behaves
nicely, although some interface work may still need to be done,
especially for the special variables and the Horizon Math, which
should be more verbose.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14347