Large rework of uiWellPropSel [devel]
Made it possible to use that class with not only
a PropertyRefSelection, but also a MnemonicSelection.

Reworked the underlaying uiWellSinglePropSel class
accordingly, to work on "names" and "units", where
the "names" are either Mnemonic or PropertyRef names.
Made it very clear that the unit selection should only
be sensitive if a null UoM is set. Otherwise it is only
a display label, read-only.
That has always been a shortcut to allow specifying a UoM
when the underlying object in the database has none. In some
cases, like for well modelling, we may even want to hide it,
because there is no input database object.
It is up to the user of the class to use that information to
propagate to the database if needed.