Cleanup of uiSynthParsGrp [devel]
Summary:
uiSynthParsGrp should be used by the SynthRock data extraction
GUI, which was currently not possible because of too many interactions
with uiStratSynthDisp. This revision ensures both class work in a much
more independent way, so that uiSynthParsGrp can be used also outside
of the SynthRock main window.
While at it, I am cleaning up its GUI and adding options to save and
restore the parameters, which is why I added a new tiny translator
and its group.

Test Plan:
Works well within uiWellExampleExtracterGrp. One or two
memory issue to be solved in uiStratSynthDisp.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14380