Improvements to StratSynth [devel]
Summary:
StratSynth::getSelection is required to be able to
restore only a subset of the synthetics stored in the object.
It avoids having to copy the object, and alter it by removing
synthetics from which ones does not want to save the parameters.

Increased the scope of StratSynth::generateOtherQuantities,
so one can set the filter using the API, and not only an
environment variable.
Also added a zstep, as most often the zstep from the post-stack
synthetic, coming from its input wavelet, is most often too large.
To get enough detail one wants to keep it low enough, with a default
value of 1ms (the data extraction GUI will have it as an extraction
parameter).

Test Plan:
saved an IOPar from the new SynthRock ML data extractio GUI.
Displayed the computed synthetics with/without filters, they remain
perfectly aligned with the synthetic seismic even though they are on a
lower sampling rate (1ms vs 4ms in my example).

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14391