Added tools for restoring the python activator in batch programs
Summary:
Setting the python activation mecanism is reserved for the GUI progs,
hence not set by batch programs. This patch makes it easy to append the required
filepaths to the command line when submitting a batch program, and ensure that
any batch program launched with theses flags will set the python activator app
as part of its initialization.
Only the call to JobSpec::setDefaultArgs() is required from the user of the class (in plugins)

[Jira: OD-3475]

Test Plan:
command line is submitted with the additional arguments.
Testing on the spawned batch process in ongoing.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14392