Fixed data corruption while reading well logs
Summary:
logs written by the od6.7 branch do not write the
log value range key in the log header.
Hence when they are read by od6.6/main, it triggers the writing
of that log data, and not all logs get written.
Since there is not yet a proper indexing of the log in the OpendTect
database, the wrong file got written.

Does not impact the public 6.6.4, but show stopper for the 6.6.5 release
This patch is only a temporary solution while a more robust indexing is implemented.
[Jira: OD-3482, OD-3364]

Test Plan:
Opened the well manager (and/or well table). Did a plot of a log
that does not have a stored value range, used the refresh button of the well
manager -> no more data corruption, and the file for that log gets updated
in content without filename change. Created a new log with the log tool,
and that log is displayed correctly in both well manager and table.
Removed that log, correctly updated in both well manager and table.
File changes in the WellInfo folder look correct during intermediate checks.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14398