Preserving the user pythonpath value
Summary:
Fixes a cascading side effect in the python activation
mecanism:
1- An OpendTect binary is started (od_main/od_DeepLearning_CC),
it initializes Basic and sets PYTHONPATH to add paths to odpy, dgbpy,
which may be in different folders especially in debug environmnents
2- That OD launches a python command (cmd prompts, bokeh training, ...)
On window this is indirectly launched using the python_activator app, to
set the python environment as active.
3- The python_activator is running is license check, which includes loading
the OpendTect Basic library at runtime. Since Basic is loaded within the
python_activator app, it sets again PYTHONPATH, overwriting any previous
value. In development environment that makes it loose the access to dgbpy,

This example shows that we have to add the paths to odpy, dgbpy on top of
the incoming environment, not as a replacement.

Can you please check https://phabricator.dgbes.com:8443/D13259,
and let me know if you think that this fix can have undesired,
difficult to check side effects.

[Jira: OD-3488]

Test Plan:
With this change I was able to run the bokeh app and run a training within
visual studio

Reviewers: wayne

Reviewed By: wayne

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14401