Fixed missing mnemonic attribute for newly created logs [devel]
Not all created logs had a mnemonic, this patch fixes it.
The usage of Well::Log::guessMnemonic should get as limited as
possible, as it should be constrained by the GUI whenever possible.
It is thus utilized in all old GUIs that do not have an
interface for showing and setting it.
One important point: Logs from the database which do not have
a mnemonic label should never receive one if they are written
back to the database. Hence we are not setting the mnemonic
in the Well::Log constructor or anywhere like this within the class.