Added tools for restoring the python activator in batch programs
SUMMARY
Setting the python activation mechanism is reserved for the GUI progs,
hence not set by batch programs. This patch makes it easy to append the required
filepaths to the command line when submitting a batch program, and ensure that
any batch program launched with theses flags will set the python activator app
as part of its initialization.
Only the call to JobSpec::setDefaultPythonArgs() is required from the user of the class (in plugins)

[Jira: OD-3475]

TEST PLAN
command line is submitted with the additional arguments.