Added few utilities to the CommandLauncher
SUMMARY
added tools that make use of the QProcess functions to redirect stdout,
stderr to files, so it does not need to be done by the called application.
Made few changes in PythonAccess in how terminal windows are started, removing the
need to use the windows "Start" command.
Also made the python activation path a positional argument, as it should be.

TEST PLAN
used from plugins. Windows terminals (the three types) start without
error, with or without the python activator app.
Background work for OD-3336.