Added few utilities to the CommandLauncher
Summary:
added tools that make use of the QProcess functions to redirect stdout,
stderr to files, so it does not need to be done by the called application.
Made few changes in PythonAccess in how terminal windows are started, removing the need to use the windows "Start" command.
Also made the python activation path a positional argument, as it should be.

Used the HiddenParam in a slightly unconventional manner, as the
default constructor is inline, and I cannot add a destructor. Should be fine
since the usage will be very limited

Test Plan:
used from plugins. Windows terminals (the three types) start without
error, with or without the python activator app.
Background work for OD-3336.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14423