Fixed cmake setup with system hdf5 libraries [devel]
On Linux, especially OpenSuse, the hdf5-devel package may be installed,
but not the hdf5-devel-static package. For that configuration the
hdf5 cmake wrapper may not be able to compile the simple test
program, because of the missing static libraries. As a result the
required cmake variables may not be set.

This patch unsets HDF5 if no cmake variable at all is found, allowing
to build without HDF5, and adds translators on Linux if some compatible
libraries are found. The test program test_hdf5 then runs with the
system installation 1.10.7 that comes with OpenSuse-15.3