Added getISODateTimeString and getLocalDateTimeFromString
Summary:
Needed for FileSystemAccess, but also better to store the iso utc date in
omf files and iopar files.
When showing a date/time to user, it should show a local readable string.

Test Plan:
Created survey and imported seismic data. Checked all dates in files and in
the managers. Looks good.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14425