Various fixes in limitTo functions [devel]
Summary:
Virtualized the limitTo function in ranges.h. Existing implementations in TrcKey(Z)Sampling classes were happily shifting the sample distributions, and/or changing the step, which is not the purpose of these functions. I have some fears that this fix will introduce few bugs because of bad usage, like in SurveyInfo::setRange where the limitTo function was used to initialize the TrcKeyZSampling, but let's figure out in due time. Fixes OD-1918 (together with another bug fix for OD-1972).

Test Plan: Processed a pickset-based ML prediction. The Zstep of the Volume subselection in the IOPar is now
correctly set to the Zstep of the input dataset.
The test program test_trckeyzsampling succeeds with the additional test samples from od6.7

Ported D12373 from od6.7