Fix for well display in Basemap after loading in 3D scene
Summary: If a well is first loaded in 3D scene (INF not loaded) and then added to the Basemap, the Well Info returns 'isLoaded=true' as the replacement velocity is updated by D2T model. But in fact Info is not loaded, which means no surfacecoord and a no-show in Basemap.

Test Plan: Loaded wells in 3D scene and Basemap with/without sync. Wells always show up.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14432