Update of ElasticPropertyRef and related classes
Summary:
The ElasticPropSelection is currently not capable of providing
a decent selection based on a list of properties or log names. The only
formula that it can safely assign is Castagna's equation, because it is hard-coded
Decided that the classes RockPhysics::Formula and ElasticFormula should derive
from Math::Formula, to have some consistency between them, and provide safe tools
for getting a form from either RockPhysics::FormulaSet or ElasticFormulaRepository,
based on a required output and input mnemonic type.

Test Plan:
code not entirely written, except for the three base classes mentionned
above. For early comments only.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14443