Jira OD-3550: Added some nullptr checks[devel]
Summary:
This revivion is in response to a crash potentially caused due to missing nullptr checks. I have added as many as I think could have helped.
The Jira issue for this crash is OD-3550.

Test Plan: nothing tested as of yet, dump file is not loading neither is there a workflow to replicate the crash

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14478