default log name also changes when renaming a default log now[devel]
Summary: Default logs are stored in a separate .defs file. This meant that when a log was renamed, the default file should also have been edited simultaneously to mirror a change in logname. This was not happening and this revision fizxes this.

Test Plan: tested in Well manager and the the default log file stored on disk

Reviewers: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14528