Fixed management of uiTaskRunner in uiStratSynthDisp [devel]
The StratSynth class does not manage its TaskRunner, but a new object
was given to it, managed by Qt (as it can be a uiDialog), and not set to
null when Qt has deleted that dialog.
Thus added the necessary memory management.
Strangely this uiDialog is deleted by Qt every time a new model is opened when
using the command driver, but not when doing it 'manually'.

[Jira: OD-3659]