Rewrote setup.Python for 1.2.8 and 1.4.0 Py envs
Summary: This setup.Python is ready for the 1.2.8 and 1.4.0 Python environments and needs to be included in the miniconda3 zip-file

Test Plan: Tested on CentOS 7 VM that does not have the original disk d24 mounted. Checked for d24/ occurences and for some of the files whether the correct path had been written after executing of script.

Reviewers: arnaud, nageswara

Reviewed By: nageswara

Subscribers: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D14601