Added tool to run python scripts installed with pip [devel]
Summary:
Python executables installed within a conda environment
with pip may not run without using the full path to the executable
as argument of the python executable, if the python environment
was relocated. Many python cache files seem to keep the original
full path to python within the environments.

Test Plan:
safety package was installed with pip inside the cuda10
environment (v1.2.8), but with conda in the new mkl and cuda11
environments (v1.4.0).
Now the safety check button works in all three environments.

Reviewers: nanne

Reviewed By: nanne

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14610