Global refPtr/unRefPtr are strictly reserved for ReferencedObject impl [devel]
Summary:
Removing the dynamic cast in these global functions, to only make them usable
by classed which are derived from ReferencedObject and no longer by their macro
equivalent.
Advantages are that it makes the usage of these derived implementations much safer,
and it allows the forward declaration of objects stored under a RefMan, instead of having
to include the corresponding header file.

As a convenience to existing classes I have added static functions to the macro, to
facilitate the migration of existing code. The only remaining disadvantage is that
storing a macro-based referenced object in a RefMan does nothing.

Test Plan: OD compiles and starts. The test program test_refcount is updated and successfull.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14644