Large update of Synthetic seismic generation [devel]
Summary:
To be able to directly generate elastic impedance
gathers, the following changes were required:

1- One needs a container that stores a set of time-depth model
for each position: The new class ReflectivityModelSet.
Will be used by both zero-offset stacks,
offset-based gathers and elastic-based stacks.
2- This container must be set exclusively from an ElasticModel,
before the RayTracer is used. Hence moved the code that computes
the time-depth models for each offset outside of the RayTracer1D.
3- This container should also have storage for the actual
reflectivities (spikes), regardless of how they are computed.
Classes like the RayTracer1D, or similar ones for
elastic/HTI reflectivities, will compute and write the reflectivities,
but stop taking care of the container data allocation (already done then).
4- The reflectivity sampler had part of the code for computing and
sorting the time sampled reflectivities, the other half being on the dgb side.
Centralized the code to ensure that this task can output, at request,
frequency and/or time-sampled reflectivities, and made a test program for it.
5- Huge update of the classes in synthseis.h/cc:
Reverted entirely the data allocation: The only class that remains exported
is the Seis::RaySynthGenerator. Should be used to compute synthetics from the
scratch or using the results of a previous run. That means that is reuses
existing timedepth models and reflectivities.
This class ensures that all traces are allocated
with the right size, and provides the data containers that will be used
during the computation by MultiTraceSynthGenerator and SynthGenerator.
Added a container, the SynthGenDataPack class, to store everything output
by Seis::RaySynthGenerator, with the exception of the seismic traces.
6- The output of any Seis::RaySynthGenerator can easily be transferred to
a new SyntheticData object by using the new static SyntheticData::get function.
All data (TD models, reflectivities and SeisTrcs) get transferred without
any copy/duplication.
7- Used this new functionality to update the storage of synthetic data and
reflectivities in the well tie module.
If can thus thereafter be used for pre-stack synthetic or elastic-based stacks without API change.
8- Updated the StratSynth class to make use of the new functionality. Since all data is
now refcounted and inside the SyntheticData objects, the RayModelManager can be
removed completely.

[Jira: OD-3373 OD-3374 OD3375 OD-3376 OD-3377]

Test Plan:
OD compiles and starts. All test programs return succeess.
Tested the well tie module in detail, StratSynth part to be tested.
Will need to check what to do with uiSynthSeis, as I am missing here a
factory for the interface uiSynthSeisGrp

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14653