Minor fixes for Color Table import
Summary:
ColorTable import looked for only the 'Name' key to list the sequences in the import window, without checking on the return value of the ColTab::Sequence::usePar function. The usePar function on the other hand returned false when Marker Color or Undef Color was not set.
Now, made the Marker Color and Undef Color parameters optional and made the import class check on the return value of usePar. Also fixed a small memory leak.

Test Plan: Imported a color table with and without the optional keys. Both results are identical.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14661