Fixed safety check not working on windows [devel]
Conda-based installations of safety check could not be executed using
the python executable and full path to the script executable, because
the encoding is not defined.
Two options can be done with conda-based scripts on windows:
1- They are executed directly, not using the python.exe application
2- There is a script file with the same basename as the script name,
and the suffix "-script.py" that can be executed as argument to
python.exe. This script sets the encoding before calling the executable.

Preferred implementing option -1-, by using the presence of the [scriptnm]-script.py
file to determine if the script was installed by conda or pip.

[Jira: OD-3750]