Fix for ColTab restore during session restore
Summary: See OD-3758. Some objects in the session file store only the sequence name if it is a System coltab. Now such coltabs will be resored from the SequenceManager. Also, if a coltab is empty and is not present in SequenceManager, the usePar function will now return false.

Test Plan: saved and restored sessions with system, edited and user coltabs. All restore fine now.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14678