Fixed python commands which stop working after switching surveys
Summary

there was a discrepency in the python activation path
saved as a class member in OD::PythonAccess between free and licensed
surveys. In the first case the full path to the conda activation script
was saved, in the second case only the full path to the python installation
(shorter then).
Made the second case equal to the first, even if licensed surveys won't
execute that script, or even need the file to the be present, because
OD will be using its own activation app, to ensure that the filepath
always has the same number of levels.

[Jira: OD-3750]

Test Plan

Applied the steps listed in the Jira issue, ensured that the
list of installed packages still matches the output of 'conda list', and
not 'pip list', and that the Machine Learning plugin succeeds in finding
its dependencies at startup.