Using object-specific font for computing height/width [devel]
Summary: Functions like setPrefWidthInChar or setPrefHeightInChar gave wrong results as a default QFont was always used for QFontMetrics. Now, if the uiObjectBody has a pre-set uiFont, then this is the one used by QFontMetrics.

Test Plan: uiTextEdit and uiTextBrowser get better (not perfect yet) initial width and height when set in chars.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14686