Moved SynthRock-related data preparation to a dedicated task
Summary:
Moving a large part of StratPropSyntheticDataCreator::doPrepare to a
dedicated parallel task to gain efficiency in the process.

[Jira: OD-3774]

Test Plan:
Restored a set of 7000 pseudowells from a simulation saved by Hardeep.
Processing time for each part was about 9sec in release mode, now the StratSeqSplitter
takes about 3sec, on 8 CPUs.
No change in the quantitative output or memory consumption.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14694