Large update of the synthetic generation GUI [devel]
Summary:
Large cleanup of uiRayTracer1D, to remove all parameters
that should be dependent on the factory implementations, and
regroup most interface items under an advanced button. Each
factory implementation can derive from uiRayTracerAdvancedGrp to
add its parameters to the interface, and such uiGroups can be directly
used in other groups, like it is done for the prestack angle mute
processing which has its own advanced button.
Made it such that the "generic" implementation is never displayed unless it
is the only item present in the factory.

Added uiSynthSeis on the same model, to provide an interface for
the Seis::SynthGenerator factory, moving some of the interface previously
in uiWellAttrib down into uiSeis.
uiSynthSeisSel should exclusively be used for a single synthetic data type,
i.e. zero offset or prestack (or elastic stack eventually), thus there will
be scope for adding a uiMultiSynthSeisSel (we need it at two locations already,
and could make use of it in the WellTie GUI as well to support elastic angle stack
very easily when they are ready).

Test Plan:
Tested GUI from SynthRock, the prestack angle mute computer and
prestack angle stacks GUI. Still need to check the actual computation for these
two when done on the fly (tests programs are OK).

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14726