Fixed a crash in ML
Summary:
Seismic was trying to retrieve Cube Data which is done via a virtual function and not a pure virtual function. The default implementation returned false and this was not being read in the calling function. This has now been corrected
Rest of the dependent change sin D14768
OD-3820

Test Plan: No crash seen

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14757