Better implementation of boundingRect()
Summary: while increasing the size of a marker or well symbol, the entire marker was not updated immediately, because the boundingRect() was based on PointSize and not pixels. Now using the screenDPI to compute the size i pixels fixes this problem.

Test Plan: Changing the size of markers and labels results in imeediate update.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14758