Fixed comparisons between strings and OD::String [devel]
C++17 enable the compiler flag /permissive- on windows
(can be found under Configuration->Properties->Language->Conformance mode.
Previous setting was "default", likely equal to /permissive+ with C++14 / MSVC2019

with /permissive-, one should not have a null terminated string and an OD::String
when using an ?: form of if-then-else statement as it triggers the compilation error C2445.
Returned values need to be both a null terminated string or both an OD::String.

see:
https://docs.microsoft.com/en-us/cpp/build/reference/permissive-standards-conformance?view=msvc-170