Fixed few issues with the MultiIDs [devel]
IOMan::get could return an object from a different group ID,
as this component was no longer checked, resulting in many
failure to retrieve the IOObj when working with different data types.

SurveyDataTreePreparer::prepDirData was not yet aware that the
main ID was swapped for custom IODirs from the groupID to the objectID,
resulting in failures to initialize both plugins GMT and Madagascar.