Added MultiID implementation for Synthetic and InMemory [devel]
Summary:
Declared two fixed objectIDs that can be utilized to distinguish
synthetic and in-memory MultiIDs.
GroupID should stay either the translator ID for database IDs,
or the DataPackMgr ID for in-memory IDs (the ones that use to start with '#'.

Added a translation in the MultiID::fromString implementation.

Test Plan:
Angle Stacks and AVO gradient attributes are now working
within SynthRock. More work required to have the Instantaneous attribute
working.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14777