Fixed loaded of well data in well log interpolation
The data preparation step was checking for well markers even if no
marker was set to be utilized, and rejected the corresponding wells.
That means that wells without any marker loaded could never be used
by the well log interpolator.
Now the data preparation step only explicitely looks for the markers
which are set as top/bottom markers in the params_ class member, if the
Z sub-selelection is marker based.

[Jira: OD-3854]