Fixed loaded of well data in well log interpolation
Summary:
The data preparation step was checking for well markers even if no
marker was set to be utilized, and rejected the corresponding wells.
That means that wells without any marker loaded could never be used
by the well log interpolator.
Now the data preparation step only explicitely looks for the markers
which are set as top/bottom markers in the params_ class member, if the
Z sub-selelection is marker based.
Also updated the code to make use of the standard API for reading well
data from its manager.

[Jira: OD-3854]

Test Plan:
processed a well log interpolation where the marker selection
is set to either Start of Data/End of Data (void) or MFS11/FS2 in F3.
The job started in both cases. Processing is successful on-the-fly and
in batch.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14806