Fixed incorrect setting of seed in random number generator
Summary:
the CEEMD attribute is setting the random number generator
seed to 0 before drawing each noise value, resulting in identical values
being drawn on Unix (not on Windows because windows ignores the seed).
That incorrect usage was not correctly protected.

[Jira: OD-3859]

Test Plan:
different values are now drawned in the CEEMD attribute on Unix
Made a test program, which actually showed many shortcoming that will be
addressed in the main branch only.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14818